requirements.txt
numpy==1.24.4
Pillow==10.1.0
tensorflow==2.12.0